deletion fix

+++ backend/models/issues.py
@@ -27,14 +27,15 @@ class Attachment(db.Model):
         return "{}/{}".format(head, thumb_name)
 
     def delete(self):
-        delete_file(self.image_url)
-        delete_file(self.get_thumbnail_url())
         db.session.delete(self)
         db.session.commit()
-        directory_path = os.path.abspath(os.path.join(
-            current_app.config['MEDIA_FOLDER'], self.image_url, os.pardir))
-        if not os.listdir(directory_path):
-            os.rmdir(directory_path)
+        if os.path.exists(self.image_url) and os.path.exists(self.get_thumbnail_url()):
+            delete_file(self.image_url)
+            delete_file(self.get_thumbnail_url())
+            directory_path = os.path.abspath(os.path.join(
+                current_app.config['MEDIA_FOLDER'], self.image_url, os.pardir))
+            if not os.listdir(directory_path):
+                os.rmdir(directory_path)
 
     def get_full_thumbnail_url(self):
         url = self.get_thumbnail_url()
